Extract transport agnostic zipkin code into finagle-zipkin-core
Move most of finagle-zipkin to finagle-zipkin-core

Pants build files for finagle-zipkin-core

Add finagle-zipkin-core as a dependency of finagle-zipkin

Separate scribe specific code

Fix non-exhaustive case warning

Move SamplingTracer to finagle-zipkin-core

Generic tracer cache with flush on shutdown

Remove scribe specific tests already in finagle-zipkin and add test for flushing tracers

Fix import

Weed out sbt and pants dependencies

Make internal classes private

Updated CHANGES, change still needs a RB_ID

Improved scaladoc comments

Annotate return types where needed

Remove confusing SamplingTracer from finagle-zipkin

Minor dependency fix

Fixes after code review

Fix some more IntelliJ formatting fuckups

Fix after failed rebase
