+++ finagle-zipkin-core/BUILD
@@ -1,0 +1,11 @@
+target(name='finagle-zipkin-core',
+  dependencies=[
+    'finagle/finagle-zipkin-core/src/main/scala'
+  ]
+)
+
+target(name='tests',
+  dependencies=[
+    'finagle/finagle-zipkin-core/src/test/scala'
+  ]
+)
