+++ finagle-zipkin-core/src/main/scala/com/twitter/finagle/zipkin/core/Span.scala
@@ -1,4 +1,4 @@
-package com.twitter.finagle.zipkin.thrift
+package com.twitter.finagle.zipkin.core
 
 /**
  * The `Span` is the core datastructure in RPC tracing. It denotes the
@@ -36,7 +36,7 @@ case class Span(
   /**
    * @return a pretty string for this span ID.
    */
-  def idString = {
+  def idString: String = {
     val spanString = traceId.spanId.toString
     val parentSpanString = traceId._parentId map (_.toString)
