+++ finagle-zipkin/src/main/resources/META-INF/services/com.twitter.finagle.tracing.Tracer
@@ -1,1 +1,1 @@
-com.twitter.finagle.zipkin.thrift.SamplingTracer
+com.twitter.finagle.zipkin.thrift.ScribeZipkinTracer
